[AMD-SMI] Fix trivial correctness and safety bugs from code hygiene audit
Fixes 10 bugs identified in the code hygiene audit (issue #3634):

- py-interface: sys.getsizeof → ctypes.sizeof for correct C type sizes
- py-interface: PAGE_FAULT_START mapped to wrong enum (copy-paste bug)
- amd_smi_no_drm_nic.cc: sizeof(array-1) → sizeof(array) - 1 in snprintf
- amd_smi_system.cc: OOB write via sockets[0] on empty vector → push_back
- amd_smi_utils.cc: num_pages = 0 assigned to pointer, not *num_pages
- rocm_smi_device.cc: isRestartInProgress checked stale input instead of
  local isSystemAMDGPUModuleLive
- rocm_smi_utils.cc: dangling pointer from local string c_str() → literal
- rocm_smi_utils.cc: pclose(NULL) UB when popen fails → add NULL guard
- amd_smi_lspci_commands.cc: multi-char literal ']----' → string "]----"
- rocm_smi_kfd.cc: stoi truncates uint64_t GPU IDs → stoull

+++ projects/amdsmi/rocm_smi/src/rocm_smi_kfd.cc
@@ -285,7 +285,7 @@ static int ReadKFDGpuId(uint32_t kfd_node_id, uint64_t *gpu_id) {
     return ENXIO;
   }
 
-  *gpu_id = static_cast<uint64_t>(std::stoi(gpu_id_str));
+  *gpu_id = std::stoull(gpu_id_str);
   return 0;
 }
